rename SubmissionGallery > GalleryManager
