+++ src/app/GalleryManager.jsx
