+++ src/index.js
@@ -26,7 +26,7 @@ import SearchEditor from 'app/SearchEditor';
 import NoMatch from 'app/NoMatch';
 import About from 'app/About';
 import SubmissionList from 'app/SubmissionList';
-import SubmissionGallery from 'app/SubmissionGallery';
+import GalleryManager from 'app/GalleryManager';
 
 // Utils
 //import registerServiceWorker from 'serviceworker!./sw.js';
@@ -111,7 +111,7 @@ class Root extends React.Component {
             <Route path="forms/create" component={FormCreate}/>
             <Route path="forms/:id" component={FormEdit}/>
             <Route path="forms/:id/submissions" component={SubmissionList}/>
-            <Route path="forms/:id/gallery" component={SubmissionGallery}/>
+            <Route path="forms/:id/gallery" component={GalleryManager}/>
           </div>
         ) : null}
         <Route path="*" component={NoMatch} />
